create MapView from code
