+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/map/MapboxMapViewBinder.kt
@@ -1,24 +1,16 @@
 package com.mapbox.navigation.dropin.map
 
-import android.view.ViewGroup
-import androidx.transition.Scene
+import android.content.Context
 import com.mapbox.maps.MapView
 import com.mapbox.maps.plugin.compass.compass
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
-import com.mapbox.navigation.dropin.R
-import com.mapbox.navigation.dropin.databinding.MapboxMapviewLayoutBinding
 
 @ExperimentalPreviewMapboxNavigationAPI
 internal class MapboxMapViewBinder : MapViewBinder() {
 
-    override fun onCreateMapView(viewGroup: ViewGroup): MapView {
-        Scene.getSceneForLayout(
-            viewGroup,
-            R.layout.mapbox_mapview_layout,
-            viewGroup.context,
-        ).enter()
-        return MapboxMapviewLayoutBinding.bind(viewGroup).mapView.also {
-            it.compass.enabled = false
+    override fun getMapView(context: Context): MapView {
+        return MapView(context).apply {
+            compass.enabled = false
         }
     }
 }
