+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/customnavview/CustomMapViewBinder.kt
@@ -1,18 +1,16 @@
 package com.mapbox.navigation.qa_test_app.view.customnavview
 
 import android.content.Context
-import android.view.ViewGroup
 import com.mapbox.maps.MapView
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.dropin.map.MapViewBinder
 
 @OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
-class CustomMapViewBinder(private val context: Context) : MapViewBinder() {
+class CustomMapViewBinder : MapViewBinder() {
 
-    override fun onCreateMapView(viewGroup: ViewGroup): MapView {
-        return customMapViewFromCode(context).also {
-            viewGroup.removeAllViews()
-            viewGroup.addView(it)
-        }
+    override fun getMapView(context: Context): MapView {
+        return customMapViewFromCode(context)
     }
+
+    override val shouldLoadMapStyle: Boolean = false
 }
